chore: Update global.scss and add _breakpoints.scss
This commit updates the global.scss file to include the CSS property `margin: 0;` for the body element and sets the line-height of anchor elements to 1. Additionally, a new file _breakpoints.scss is added, defining the `$table-screen-size` variable with a value of 960px. These changes improve the overall styling and responsiveness of the application.