chore: Update max-width in page.module.scss and header/index.module.scss
This commit updates the max-width property in the page.module.scss and header/index.module.scss files. In page.module.scss, the max-width is changed to calc(breakpoints.$lg + 4rem) to compensate for padding. In header/index.module.scss, the max-width is removed. These changes improve the layout and responsiveness of the page and header components.